requirements.txt
# Umgebungs-Loader
python-dotenv>=1.0.0

# OpenAI & GROQ Clients
openai>=0.27.0
groq>=0.1.0

# RAG & Text-Splitting
langchain>=0.0.300

# Vektor-Datenbanken / Indexe
faiss-cpu>=1.7.4
qdrant-client>=1.5.0   # optional
pinecone-client>=2.2.1 # optional

# Reranking
sentence-transformers>=2.2.2

# PDF-Parsing
pdfplumber>=0.9.0
PyPDF2>=3.0.1

# Frontend
streamlit>=1.25.0

# Testing
pytest>=7.2.0

openai
groq
python-dotenv
pdfplumber
langchain
faiss-cpu
sentence-transformers
streamlit
pytest

faiss-cpu
sentence-transformers
pdfplumber
python-dotenv